requirements.txt
Django==1.9.2
PyJWT==1.4.0
Wand==0.4.2
Willow==0.2.2
django-taggit==0.18.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
six==1.10.0
